+++ plugins/cicd-statistics/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-cicd-statistics
 
+## 0.1.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.8.2
+  - @backstage/plugin-catalog-react@1.9.3
+  - @backstage/catalog-model@1.4.3
+
 ## 0.1.31-next.2
 
 ### Patch Changes