+++ plugins/entity-feedback-backend/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-entity-feedback-backend
 
+## 0.2.6
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/backend-common@0.20.1
+  - @backstage/catalog-client@1.5.2
+  - @backstage/backend-plugin-api@0.6.9
+  - @backstage/plugin-auth-node@0.4.3
+  - @backstage/catalog-model@1.4.3
+  - @backstage/config@1.1.1
+  - @backstage/plugin-entity-feedback-common@0.1.3
+
 ## 0.2.6-next.2
 
 ### Patch Changes