+++ plugins/firehydrant/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-firehydrant
 
+## 0.2.13
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/core-components@0.13.10
+  - @backstage/core-plugin-api@1.8.2
+  - @backstage/plugin-catalog-react@1.9.3
+  - @backstage/catalog-model@1.4.3
+
 ## 0.2.13-next.2
 
 ### Patch Changes