+++ plugins/lighthouse-backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-lighthouse-backend
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.20.1
+  - @backstage/catalog-client@1.5.2
+  - @backstage/backend-plugin-api@0.6.9
+  - @backstage/plugin-catalog-node@1.6.1
+  - @backstage/backend-tasks@0.5.14
+  - @backstage/catalog-model@1.4.3
+  - @backstage/config@1.1.1
+  - @backstage/types@1.1.1
+  - @backstage/plugin-lighthouse-common@0.1.4
+
 ## 0.4.1-next.2
 
 ### Patch Changes