+++ plugins/scaffolder-backend-module-gitlab/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-scaffolder-backend-module-gitlab
 
+## 0.2.12
+
+### Patch Changes
+
+- 604c9dd: Add Scaffolder custom action that creates GitLab issues called `gitlab:issues:create`
+- 7c522c5: Add `gitlab:repo:push` scaffolder action to push files to arbitrary branch without creating a Merge Request
+- Updated dependencies
+  - @backstage/backend-common@0.20.1
+  - @backstage/plugin-scaffolder-node@0.2.10
+  - @backstage/config@1.1.1
+  - @backstage/errors@1.2.3
+  - @backstage/integration@1.8.0
+
 ## 0.2.12-next.2
 
 ### Patch Changes